% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_to_useful.R
\name{add_C_mats}
\alias{add_C_mats}
\title{Add allocation matrices to a data frame}
\usage{
add_C_mats(completed_allocation_tables)
}
\arguments{
\item{completed_allocaiton_tables}{}
}
\value{
A data frame with a \code{C} column containing allocation matrices.
}
\description{
This function adds allocation (\code{C}) matrices to the previously-created
\code{CompletedAllocationTables} target.
}
