% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregate_products}
\alias{aggregate_products}
\title{Aggregate by product}
\usage{
aggregate_products(
  .psut_df,
  matrices_to_aggregate = c("R", "U", "V", "Y"),
  aggregation_map = PFUWorkflow::product_aggregation_map,
  countries,
  country = IEATools::iea_cols$country
)
}
\arguments{
\item{.psut_df}{A data frame with PSUT matrices in columns.}

\item{matrices_to_aggregate}{A list of the matrices to aggregate according to \code{aggregation_map}.
Default is \code{c("R", "U", "V", "Y")}.}

\item{aggregation_map}{A description of the aggregation to be performed.
See \code{matsbyname::aggregate_byname()} for a description of its format.
Default is \code{PFUWorkflow::product_aggregation_map}.}

\item{countries}{The countries to be analyzed.}

\item{country}{The name of the country column in \code{.psut_df}.}
}
\value{
A data frame of aggregated products.
}
\description{
This function aggregates PSUT matrices by energy carrier (product)
according to an aggregation map.
Aggregated matrices replace original matrices on output.
}
