% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{alloc_plots_df}
\alias{alloc_plots_df}
\title{Create allocation graphs in a data frame}
\usage{
alloc_plots_df(
  completed_alloc_tab_targ = SEAPSUTWorkflow::target_names$CompletedAllocationTables,
  cache_path = ".drake",
  .df = drake::readd(completed_alloc_tab_targ, path = cache_path, character_only =
    TRUE),
  plots = "plots",
  country = IEATools::iea_cols$country,
  ef_product = IEATools::template_cols$ef_product,
  destination = IEATools::template_cols$destination,
  year = IEATools::iea_cols$year,
  .values = IEATools::template_cols$.values,
  machine = IEATools::template_cols$machine,
  eu_product = IEATools::template_cols$eu_product
)
}
\arguments{
\item{completed_alloc_tab_targ}{The string name of the completed allocation tables target. Default is \code{SEAPSUTWorkflow::target_names$CompletedAllocationTables}.}

\item{cache_path}{The path to the drake cache from which \code{completed_allocation_tables_target} will be read. Default is ".drake".}

\item{.df}{The completed allocation tables data frame. Default is \code{drake::readd(completed_allocation_tables_target, path = cache_path, character_only = TRUE)}.}

\item{plots}{The name of the output column containing allocation graphs. Default is "plots".}

\item{country}{See \code{IEATools::iea_cols}.}

\item{ef_product, destination}{See \code{IEATools::template_cols}.}

\item{year}{See \code{IEATools::iea_cols}. Passed to \code{alloc_graph()}.}

\item{.values, machine, eu_product}{See \code{IEATools::template_cols}. Passed to \code{alloc_graph()}.}
}
\value{
A data frame containing a list column of \code{ggplot2} allocation graphs.
}
\description{
This function adds a column of \code{ggplot2} graphs to a completed allocation tables data frame.
The graphs are stored in a list column named \code{plots}.
}
\details{
By default, the completed allocation tables data frame is read from a \code{drake} cache.

The data frame is grouped by \code{country}, \code{ef_product}, and \code{destination} and nested prior to making the graphs,
meaning that one allocation graph is constructed for each combination of country, ef_product, and destination.
}
\examples{
# Make a simple data frame with the expected structure.
alloc_table <- tibble::tribble(~Country, ~Year, ~Ef.product, ~Destination,
                               ~.values, ~Machine, ~Eu.product,
                               "GHA", 1971, "Gasoline", "Transport",
                               0.5, "Cars", "MD",
                               "GHA", 1971, "Gasoline", "Transport",
                               0.5, "Trucks", "MD",
                               "GHA", 2020, "Gasoline", "Transport",
                               0.2, "Cars", "MD",
                               "GHA", 2020, "Gasoline", "Transport",
                               0.8, "Trucks", "MD",
                               "ZAF", 1971, "Gasoline", "Transport",
                               0.5, "Cars", "MD",
                               "ZAF", 1971, "Gasoline", "Transport",
                               0.5, "Trucks", "MD",
                               "ZAF", 2020, "Gasoline", "Transport",
                               0.3, "Cars", "MD",
                               "ZAF", 2020, "Gasoline", "Transport",
                               0.7, "Trucks", "MD")
alloc_plots_df(.df = alloc_table)
}
