% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{assemble_eta_fu_tables}
\alias{assemble_eta_fu_tables}
\title{Assemble completed final-to-useful efficiency tables}
\usage{
assemble_eta_fu_tables(
  incomplete_eta_fu_tables,
  exemplar_lists,
  completed_fu_allocation_tables,
  countries,
  which_quantity = c(IEATools::template_cols$eta_fu, IEATools::template_cols$phi_u),
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  exemplars = SEAPSUTWorkflow::exemplar_names$exemplars,
  exemplar_tables = SEAPSUTWorkflow::exemplar_names$exemplar_tables,
  alloc_data = SEAPSUTWorkflow::exemplar_names$alloc_data,
  incomplete_eta_tables = SEAPSUTWorkflow::exemplar_names$incomplete_eta_table,
  complete_eta_tables = SEAPSUTWorkflow::exemplar_names$complete_eta_table
)
}
\arguments{
\item{countries}{}
}
\value{

}
\description{
Assemble completed final-to-useful efficiency tables
}
