% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{assemble_fu_allocation_tables}
\alias{assemble_fu_allocation_tables}
\title{Assemble completed final-to-useful allocation tables}
\usage{
assemble_fu_allocation_tables(
  incomplete_allocation_tables,
  exemplar_lists,
  specified_iea_data,
  countries
)
}
\arguments{
\item{exemplar_lists_target}{}
}
\value{

}
\description{
This function is used in a drake workflow to assemble completed final-to-useful allcoation tables.
}
