% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{assemble_phi_u_tables}
\alias{assemble_phi_u_tables}
\title{Assemble completed phi (exergy-to-energy ratio) tables}
\usage{
assemble_phi_u_tables(
  incomplete_phi_u_table,
  phi_constants_table,
  countries,
  max_year = NULL,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  .values = IEATools::template_cols$.values,
  eu_product = IEATools::template_cols$eu_product,
  product = IEATools::iea_cols$product,
  phi_colname = IEATools::phi_constants_names$phi_colname,
  phi_source_colname = IEATools::phi_constants_names$phi_source_colname,
  is_useful = IEATools::phi_constants_names$is_useful_colname,
  eta_fu_tables = SEAPSUTWorkflow::phi_source_options$eta_fu_tables,
  phi_constants = SEAPSUTWorkflow::phi_source_options$phi_constnats
)
}
\arguments{
\item{phi_constants_table}{A data frame of constant phi values with reasonable default values for all energy products.}

\item{countries}{A vector of countries for which completed phi tables are to be assembled.}

\item{max_year}{The latest year for which analysis is desired. Default is \code{NULL}, meaning analyze all years.}

\item{incomplete_phi__utable}{A data frame of phi values read from machine efficiency and phi data files.
This data frame can be "incomplete," i.e., it can be missing
phi values.
The phi values from \code{phi_constants_table} will be used instead.}
}
\value{
A data frame of phi values for every country, year, machine, destination, etc.
}
\description{
This function is used in the drake workflow to assemble completed phi (exergy-to-energy ratio) tables
given a set of phi tables read from machine data files and a phi constants table.
The algorithm gives priority in this order:
\enumerate{
\item phi values from the \code{incomplete_phi_table} argument
\item phi values from the \code{phi_constants_table} argument
}
}
