% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exemplar_table_functions.R
\name{build_one_exemplar_list}
\alias{build_one_exemplar_list}
\title{Make a list of exemplar countries}
\usage{
build_one_exemplar_list(p_names, exemp, restofworldcode, world)
}
\arguments{
\item{p_names}{A character vector of previous names for this country.}

\item{exemp}{An exemplar country.}

\item{restofworldcode}{The rest-of-world code for this country.}

\item{world}{A string indicating the world.}
}
\value{
A list of length 1 containing all possible exemplars, in the order in which they apply.
}
\description{
This is a non-public, internal helper function
}
