% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cache_info}
\alias{cache_info}
\title{Cache information}
\format{
A string list with 1 entries.
\describe{
\item{cache_path}{The default path to the drake cache.}
}
}
\usage{
cache_info
}
\description{
A string list containing information about the drake cache.
Items in the list provide default values for column name function arguments
throughout the \code{SEAPSUTWorkflow} package.
}
\examples{
cache_info
}
\keyword{datasets}
