% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_to_useful.R
\name{calc_eta_fu_phi_u_vecs}
\alias{calc_eta_fu_phi_u_vecs}
\title{Add efficiency (\code{eta}) and exergy-to-energy ratio (\code{phi}) vectors
to a data frame.}
\usage{
calc_eta_fu_phi_u_vecs(
  completed_efficiency_tables,
  completed_phi_tables,
  countries,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  c_source = IEATools::template_cols$c_source,
  eta_fu_phi_u_source = IEATools::template_cols$eta_fu_phi_u_source,
  .values = IEATools::template_cols$.values,
  eta_fu = IEATools::template_cols$eta_fu,
  phi_u = IEATools::template_cols$phi_u
)
}
\arguments{
\item{completed_efficiency_tables}{The completed efficiency tables from which allocation (\code{eta_fu}) vectors
and exergy-to-energy ratio vectors (\code{phi_u}) should be created.
This data frame is most likely to be the \code{CompletedEfficiencyTables} target.}

\item{countries}{The countries for which \code{eta_fu} and \code{phi_u} vectors should be formed.}

\item{country, year}{See \code{IEATools::ieacols}.}

\item{c_source, eta_fu_phi_u_source, .values, eta_fu, phi_u}{See \code{IEATools::template_cols}.}
}
\value{
A data frame with \code{eta_fu} and \code{phi_u} vectors added as columns.
}
\description{
This function adds final-to-useful efficiency (\code{eta}) and
exergy-to-energy ratio vectors to the previously-created \code{WithCmats} target.#'
}
