% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tes_tfc_functions.R
\name{calculate_all_ex_data}
\alias{calculate_all_ex_data}
\title{Create a data frame containing all aggregate energy/exergy data}
\usage{
calculate_all_ex_data(.sutdata, fd_sectors, p_industry_prefixes)
}
\arguments{
\item{.sutdata}{A data frame containing Physical Supply-Use Table (PSUT)
matrices with associated final demand sector names}

\item{fd_sectors}{A character vector of final demand sectors.}

\item{p_industry_prefixes}{A character vector of primary energy industry prefixes.
Usually "Resources", "Production", "Imports", and "Stock changes".}
}
\value{
A data frame containing energy/exergy values aggregated by total,
sector and product at each stage of the energy conversion chain.
}
\description{
This functions creates a single data frame containing the total energy by country,
year, method, energy quantification, stage (Primary, Final, Useful....),
and grouping variable (Total, Product, and Flow or Sector) using the functions:
\code{calculate_fu_ex_total}, \code{calculate_fu_ex_sector}, \code{calculate_fu_ex_product},
\code{calculate_p_ex_total}, \code{calculate_p_ex_flow}, \code{calculate_p_ex_product},
and binding the outputs of these functions into a single data frame.
}
\examples{
library(Recca)
all_data <- calculate_all_ex_data(.sutdata = Recca::UKEnergy2000mats,
                                  fd_sectors = c("Residential"),
                                  p_industry_prefixes = c("Resources"))

}
