% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tes_tfc_functions.R
\name{calculate_fu_ex_product}
\alias{calculate_fu_ex_product}
\title{Calculate total final demand of final and useful energy by product}
\usage{
calculate_fu_ex_product(PSUT_DF)
}
\arguments{
\item{PSUT_DF}{}
}
\value{
A data frame containing final and useful demand by total, by product
}
\description{
Calculate total final demand of final and useful energy by product
}
