% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tes_tfc_functions.R
\name{calculate_p_ex_total}
\alias{calculate_p_ex_total}
\title{Calculate total primary energy}
\usage{
calculate_p_ex_total(PSUT_DF)
}
\arguments{
\item{PSUT_DF}{}
}
\value{

}
\description{
Calculate total primary energy
}
