% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tes_tfc_functions.R
\name{calculate_primary_ex_data}
\alias{calculate_primary_ex_data}
\title{Create a data frame containing primary aggregate energy/exergy data}
\usage{
calculate_primary_ex_data(.sutdata, p_industry_prefixes)
}
\arguments{
\item{.sutdata}{A data frame containing Physical Supply-Use Table (PSUT)
matrices with associated final demand sector names}

\item{p_industry_prefixes}{A character vector of primary energy industry prefixes.
Usually "Resources", "Imports", and "Stock changes".}
}
\value{
A data frame containing primary energy/exergy values aggregated by total,
flow and product.
}
\description{
This functions creates a single data frame containing the total energy/exergy by country,
year, method, energy quantification, and grouping variable (Total, Product, and Flow),
for the Primary stage using the functions:
\code{calculate_p_ex_total}, \code{calculate_p_ex_flow}, \code{calculate_p_ex_product},
and binding the outputs of these functions into a single data frame.
}
\examples{
library(Recca)
primary_data <- Recca::UKEnergy2000mats \%>\%
                  tidyr::pivot_wider(names_from = matrix.name,
                                     values_from = matrix) \%>\%
                  dplyr::mutate(Method = "PCM") \%>\%
                  calculate_primary_ex_data(p_industry_prefixes = list(c("Resources", "Imports")))

}
