% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{clean_targets}
\alias{clean_targets}
\title{Cleans drake targets}
\usage{
clean_targets(
  to_clean = PFUWorkflow::target_names[seq(which(PFUWorkflow::target_names ==
    first_target), which(PFUWorkflow::target_names == last_target))],
  first_target = "CEDAData",
  last_target = PFUWorkflow::target_names[[length(PFUWorkflow::target_names)]],
  path = NULL
)
}
\arguments{
\item{to_clean}{A vector of target name strings to be cleaned. Default is
all target names from "CEDAData" to the end of \code{PFUWorkflow::target_names}.}

\item{first_target}{The string name of the first target to clean. Default is "CEDAData".}

\item{last_target}{The string name of the last target to clean. Default is the last target in the workflow.}

\item{path}{The path to the drake cache. Default is \code{NULL}, meaning that drake should look in the ".drake" folder for cache information.}
}
\value{
Nothing. This function should be called for its side effect of precision cleaning the drake cache.
}
\description{
It is often helpful to clean only a few targets.
This function allows specified targets to be cleaned.
By default, this function cleans all targets \emph{after} \code{IEAData},
according to the list \code{PFUWorkflow::target_names}.
}
