% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tes_tfc_functions.R
\name{create_fd_sectors}
\alias{create_fd_sectors}
\title{Create a vector of final demand sectors}
\usage{
create_fd_sectors()
}
\value{
A vector of final demand sectors
}
\description{
Create a vector of final demand sectors
}
