% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tes_tfc_functions.R
\name{create_fd_sectors_list}
\alias{create_fd_sectors_list}
\title{Create a list containing final demand sectors equal to the length of a dataframe}
\usage{
create_fd_sectors_list(fd_sectors, PSUT_DF)
}
\arguments{
\item{fd_sectors}{A vector of final demand sectors, usually produced by
\code{create_fd_sectors}}

\item{PSUT_DF}{A data frame containing PSUT matrices that require associated
final demand sector names}
}
\value{
A list the length of a desired data frame containing final demand vectors
}
\description{
Create a list containing final demand sectors equal to the length of a dataframe
}
