% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tes_tfc_functions.R
\name{create_p_industry_prefixes}
\alias{create_p_industry_prefixes}
\title{Create a list of primary industry prefixes}
\usage{
create_p_industry_prefixes()
}
\value{
A list of primary industry prefixes
}
\description{
Create a list of primary industry prefixes
}
