% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{dir.create.pipe}
\alias{dir.create.pipe}
\title{Pipe-amenable directory creation}
\usage{
dir.create.pipe(path, showWarnings = TRUE, recursive = FALSE, mode = "0777")
}
\arguments{
\item{path}{a character vector containing a single path name. Tilde expansion is done.}

\item{showWarnings}{logical; should the warnings on failure be shown?}

\item{recursive}{logical. Should elements of the path other than the last be created?
If true, like the Unix command \code{mkdir -p}.}

\item{mode}{the mode to be used on Unix-alikes}
}
\value{
The \code{path} argument.
}
\description{
This function is small wrapper on \code{dir.create()},
returning the path created.
If the directory creation process was unsuccessful, a warning is emitted.
}
\examples{
\dontrun{
  td <- tempdir()
  dir.create.pipe(td, recursive = TRUE)
  unlink(td, force = TRUE, recursive = TRUE)
}
}
