% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{eta_fu_plots_df}
\alias{eta_fu_plots_df}
\title{Create eta_fu graphs in a data frame}
\usage{
eta_fu_plots_df(
  .df,
  countries,
  plots = "plots",
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  .values = IEATools::template_cols$.values,
  machine = IEATools::template_cols$machine,
  eu_product = IEATools::template_cols$eu_product,
  machine_eu_product = machine_eu_product
)
}
\arguments{
\item{.df}{The completed allocation tables data frame. Default is \code{drake::readd(SEAPSUTWorkflow::target_names$CompletedEfficiencyTables, path = cache_path, character_only = TRUE)}.}

\item{countries}{The countries for which allocation plots are to be created.}

\item{plots}{The name of the output column containing allocation graphs. Default is "plots".}

\item{country}{See \code{IEATools::iea_cols}.
#@param ef_product,destination See \code{IEATools::template_cols}.}

\item{year}{See \code{IEATools::iea_cols}. Passed to \code{alloc_graph()}.}

\item{.values, machine, eu_product}{See \code{IEATools::template_cols}. Passed to \code{eta_fu_graph()}.}
}
\value{
A data frame containing a list column of \code{ggplot2} allocation graphs.
}
\description{
This function adds a column of \code{ggplot2} graphs to a completed eta_fu tables data frame.
The graphs are stored in a list column named \code{plots}.
}
\details{
By default, the completed eta_fu tables data frame is read from a \code{drake} cache.

The data frame is grouped by \code{country}, \code{machine}, and \code{eu_product } and nested prior to making the graphs,
meaning that one allocation graph is constructed for each combination of country, machine, and eu_product.
}
\examples{
# Make a simple data frame with the expected structure.
alloc_table <- tibble::tribble(~Country, ~Year, ~Ef.product, ~Destination,
                               ~.values, ~Machine, ~Eu.product,
                               "GHA", 1971, "Gasoline", "Transport",
                               0.5, "Cars", "MD",
                               "GHA", 1971, "Gasoline", "Transport",
                               0.5, "Trucks", "MD",
                               "GHA", 2020, "Gasoline", "Transport",
                               0.2, "Cars", "MD",
                               "GHA", 2020, "Gasoline", "Transport",
                               0.8, "Trucks", "MD",
                               "ZAF", 1971, "Gasoline", "Transport",
                               0.5, "Cars", "MD",
                               "ZAF", 1971, "Gasoline", "Transport",
                               0.5, "Trucks", "MD",
                               "ZAF", 2020, "Gasoline", "Transport",
                               0.3, "Cars", "MD",
                               "ZAF", 2020, "Gasoline", "Transport",
                               0.7, "Trucks", "MD")
alloc_plots_df(alloc_table, c("GHA", "ZAF"))
}
