% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_functions.R
\name{generate_fu_allocation_template}
\alias{generate_fu_allocation_template}
\title{Create a final-to-useful allocation template}
\usage{
generate_fu_allocation_template(
  country,
  data_target = "Specified",
  paths_target = "paths",
  fu_analysis_path_name = "fu_analysis_path",
  fu_allocation_template_file_name = paste0(country, " FU Allocations Template"),
  ext = ".xlsx",
  n_allocation_rows = 4,
  overwrite = FALSE
)
}
\arguments{
\item{country}{a string of the 3-letter ISO country code}

\item{data_target}{the target in the drake cache from which the template is created. Default is "Specified".}

\item{paths_target}{the name of the paths object in the drake cache. Default is "paths".}

\item{fu_analysis_path_name}{the name of the final-to-useful path member of the paths object. Default is "fu_analysis_path".}

\item{fu_allocation_template_file_name}{the file name for the template. Default is "\code{country} FU Allocations Template.\code{ext}".}

\item{ext}{the file name extension. Default is ".xlsx" for an Excel file.}

\item{n_allocation_rows}{the number of allocation rows per combination of sector and final energy carrier. Default is \code{4}.}

\item{overwrite}{tells whether to overwrite an existing file named \code{fu_allocation_file_name}.\code{ext} in location \code{fu_analysis_path_name}.
Default is \code{FALSE}.}
}
\value{
the path to the blank final-to-useful allocation template
}
\description{
Create a final-to-useful allocation template
}
