% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{generate_reports}
\alias{generate_reports}
\title{Generate reports found in report directory}
\usage{
generate_reports(
  report_source_files,
  report_dest_files,
  archive_reports = TRUE,
  timestamp_sep = "__"
)
}
\arguments{
\item{report_source_files}{A vector of absolute paths to reports that need to be run,
probably generated by \code{report_source_paths()}.}

\item{report_dest_files}{A vector of absolute paths into which completed reports are
to be written.}

\item{archive_reports}{A boolean that tells whether to archive reports in an adjacent folder.
Set \code{TRUE} to archive, \code{FALSE} to not archive.
Default is \code{TRUE}.}

\item{timestamp_sep}{A string that separates the date/time string from the report name for the file name on output.
Default is "__".}
}
\value{
A boolean. \code{TRUE} for success with all reports, \code{FALSE} if an error occurred with any report.
}
\description{
This function builds all \code{report_source_files} with \code{knitr::knit()}.
}
\details{
Each report name is suffixed by date and time (down to the second).
Duplicate report names in \code{report_source_files} generates an error.

Previous reports are shuffled to a folder named \verb{<<report_name>>_prev}.
}
