% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{get_plan}
\alias{get_plan}
\title{Create a drake plan for societal exergy analysis}
\usage{
get_plan(iea_data_path, countries, max_year)
}
\arguments{
\item{iea_data_path}{The path to IEA data in .csv format}

\item{countries}{A vector of country abbreviations to be analyzed, such as "c('GHA', 'ZAF')".}

\item{max_year}{The last year to be studied, typically the last year for which data are available.}
}
\value{
a drake plan object
}
\description{
Creates a drake workflow for societal exergy analysis.
The caller specifies location of IEA data,
which countries should be analyzed, and
which the maximum year to ba analyzed.
}
\examples{
get_plan(iea_data_path = "mypath", countries = c("GHA", "ZAF"), max_year = 1999)
}
