% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exemplar_functions.R
\name{load_exemplar_table}
\alias{load_exemplar_table}
\title{Read an exemplar table from an Excel file}
\usage{
load_exemplar_table(
  exemplar_table_path = sample_exemplar_table_path(),
  exemplar_table_tab_name = SEAPSUTWorkflow::exemplar_names$exemplar_tab_name,
  prev_names = SEAPSUTWorkflow::exemplar_names$prev_names,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year
)
}
\arguments{
\item{exemplar_table_path}{The path to the Excel file containing an exemplar table.
Default is the value of \code{sample_exemplar_table_path()}.}

\item{exemplar_table_tab_name, prev_names}{See \code{SEAPSUTWorkflow::exemplar_names}.}

\item{year, country}{See \code{IEATools::iea_cols}.}
}
\value{
an exemplar table
}
\description{
An exemplar table has a column of country names
and additional metadata,
including the exemplar country and any name changes through time.
}
\details{
This function gathers (actually, \code{tidyr::pivot_longer()}) years into a Year column.
It also
}
\examples{
load_exemplar_table()
}
