% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{load_fu_efficiency_tables}
\alias{load_fu_efficiency_tables}
\title{Load FU efficiency tables}
\usage{
load_fu_efficiency_tables(
  fu_analysis_folder,
  countries,
  file_suffix = " FU Analysis.xlsx",
  use_subfolders = TRUE
)
}
\arguments{
\item{fu_analysis_folder}{The folder from which final-to-useful analyses will be loaded.}

\item{countries}{The countries for which allocation tables should be loaded.}

\item{file_suffix}{The suffix for the FU analysis files. Default is " FU Analysis.xlsx".}

\item{use_subfolders}{Tells whether to look for files in subfolders named by \code{countries}.}
}
\value{
A data frame of FU efficiency tables read by \code{IEATools::load_eta_fu_data()}.
\code{NULL} if no data are found.
}
\description{
This function reads all final-to-useful efficiency data
in files in the \code{fu_analysis_folder} that start with the country prefixes
given in \code{countries}.
}
\details{
By default, it is assumed that each country's final-to-useful analysis file will be in a subfolder
of \code{fu_analysis_path}.
Set \code{use_subfolders} to \code{FALSE} to change the default behavior.
}
