% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{make_balanced}
\alias{make_balanced}
\title{Balance IEA data}
\usage{
make_balanced(
  IEAData,
  countries,
  grp_vars = c("Country", "Method", "Energy.type", "Last.stage", "Year", "Product"),
  country_colname = "Country"
)
}
\arguments{
\item{IEAData}{A tidy IEA data frame}

\item{countries}{The countries that should be balanced}

\item{grp_vars}{The groups that should be checked for energy balance.
Default is \code{c("Country", "Method", "Energy.type", "Last.stage", "Product")}.}

\item{country_colname}{The name of the country column in \code{IEAData}. Default is "Country".}
}
\value{
balanced IEA data
}
\description{
Balances the IEA data in a way that is amenable to drake subtargets.
Internally, this function uses \code{IEATools::fix_tidy_iea_df_balances()}.
Grouping is doing internal to this function using the value of \code{grp_vars}.
}
