% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{make_psut}
\alias{make_psut}
\title{Convert to PSUT matrices}
\usage{
make_psut(SpecifiedIEAData, countries)
}
\arguments{
\item{SpecifiedIEAData}{A data frame that has already been specified via \code{specify()}.}

\item{countries}{The countries you want to convert to PSUT matrices.}
}
\value{
a \code{matsindf}-style data frame
}
\description{
Converts tidy IEA data to PSUT matrices in a way that is amenable to drake subtargets.
Internally, \code{IEATools::prep_psut()} does the conversion to matrices.
}
