% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phi_sources}
\alias{phi_sources}
\title{Sources for phi (exergy-to-energy ratio) data}
\format{
A string list with 3 entries.
\describe{
\item{eta_fu_tables}{Indicates phi values came from final-to-useful efficiency tables.}
\item{temperature_data}{Indicates phi values came from national monthly temperature data.}
\item{phi_constants}{Indicates phi values came from a phi constants file.}
}
}
\usage{
phi_sources
}
\description{
A string list containing options for the source of phi (exergy-to-energy ratio) data.
}
\examples{
phi_sources
}
\keyword{datasets}
