% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{product_aggregation_map}
\alias{product_aggregation_map}
\title{A map for aggregating products}
\format{
A string list with 3 entries.
\describe{
\item{Oil and oil products}{All energy products counted as oil and oil products.}
\item{Natural gas}{All energy products counted as natural gas.}
\item{Renewables}{All energy products counted as renewables.}
\item{Biofuels and waste products}{All energy products counted as biofuels and waste products.}
}
}
\usage{
product_aggregation_map
}
\description{
A string list containing a map for aggregating energy products.
}
\examples{
product_aggregation_map
}
\keyword{datasets}
