% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_functions.R
\name{read_all_eta_files}
\alias{read_all_eta_files}
\title{Create a data frame containing machine Eta.fu and Phi.u values.}
\usage{
read_all_eta_files(
  eta_fin_paths,
  efficiency_tab_name = SEAPSUTWorkflow::machine_constants$efficiency_tab_name,
  year = IEATools::iea_cols$year,
  .values = IEATools::template_cols$.values
)
}
\arguments{
\item{eta_fin_paths}{A list of the file paths to machine excel files containing
FIN_ETA front sheets, and therefore usable data.
Created by calling the \code{get_eta_filepaths()} function.}

\item{efficiency_tab_name}{See \code{SEAPSUTWorkflow::machine_constants}.}

\item{year}{See \code{IEATools::iea_cols}.}

\item{.values}{See \code{IEATools::template_cols}.}
}
\value{
A data frame containing all Eta.fu and Phi.u values present
in all Machine excel files, with the following column names:
"Country", "Energy.type", "Last.stage", "Method", "Machine",
"Eu.product", "Quantity", "Year", "Value".
}
\description{
This function reads the files in \code{eat_fin_paths}
and creates a data frame of important efficiency and other variables.
}
\details{
Note that \code{eta_fin_paths} can be a directory, in which case
\code{get_eat_filepaths()} is called internally before
reading the files and creating the data frames.
}
