% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{readd_by_country}
\alias{readd_by_country}
\title{Read a subtarget based on country}
\usage{
readd_by_country(target, country, name_of_countries_object = "countries")
}
\arguments{
\item{target}{the name of the drake target as a string}

\item{country}{the 3-letter ISO abbreviation of the country for whom \code{target} is to be readd from the drake cache, as a string}

\item{name_of_countries_object}{a string giving the name of the countries object. Default is "countries".}
}
\value{
the country-specific version of \code{target}
}
\description{
Read a subtarget based on country
}
