% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storing.R
\name{release_psut}
\alias{release_psut}
\title{Save the \code{PSUT} target in a pinboard.}
\usage{
release_psut(workflow_releases_folder, psut, release = FALSE)
}
\arguments{
\item{workflow_releases_folder}{The folder that contains the pinboard.}

\item{psut}{The PSUT object, the final target for this workflow.}

\item{release}{A boolean telling whether to do a release.
Default is \code{FALSE}.}
}
\value{
If \code{release} is \code{TRUE},
the fully-qualified path name of the PSUT file in the pinboard.
If \code{release} is \code{FALSE},
\code{NULL}.
}
\description{
Save the \code{PSUT} target in a pinboard.
}
