% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report_dest_paths}
\alias{report_dest_paths}
\title{Create file paths for reports output}
\usage{
report_dest_paths(report_source_files, report_dest_folder)
}
\arguments{
\item{report_source_files}{A vector of report source file names, probably created with \code{report_source_paths()}.}

\item{report_dest_folder}{A directory for report output.}
}
\value{
A vector of file paths for reports output.
}
\description{
Create a vector of file paths for reports output.
}
