% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report_source_paths}
\alias{report_source_paths}
\title{List reports found in report directory}
\usage{
report_source_paths(
  report_source_folders = system.file(file.path("reports"), package =
    "SEAPSUTWorkflow"),
  report_suffixes = c(".Rmd", ".Rnw")
)
}
\arguments{
\item{report_suffixes}{A vector of strings of report suffixes. Default is \code{c(".Rmd", ".Rnw")}.}

\item{report_dirs}{A string that identifies a directory of reports to be generated.
Default is \code{system.file(file.path("reports"), package = "SEAPSUTWorkflow")},
which provides some default reports.}
}
\value{
A vector of absolute paths to reports to be generated.
}
\description{
Makes a list of raw report files in \code{report_dirs}.
This function descends the directories rooted at \code{report_dirs} to find all reports to be run.
Reports to be run are identified by filename extensions,
by default \code{c(".Rmd", ".Rnw")}.
}
\details{
Users can override the default value of \code{report_dirs} to supply their own
report directory(ies).
If users want to also include the default reports,
be sure to include \code{system.file(file.path("reports"), package = "SEAPSUTWorkflow")}
in the vector of paths supplied to \code{report_dirs}.
}
