% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{reports_to_generate}
\alias{reports_to_generate}
\title{List reports found in report directory}
\usage{
reports_to_generate(report_dir, report_suffixes = c(".Rmd", ".Rnw"))
}
\arguments{
\item{report_dir}{A string that identifies a directory of reports to be generated.}

\item{report_suffixes}{A vector of strings of report suffixes. Default is \code{c(".Rmd", ".Rnw")}.}
}
\value{
A vector of absolute paths to reports to be generated.
}
\description{
A list of raw report files in \code{report_dir}.
This function descends the directory rooted at \code{report_dir} to find all reports to be run.
Reports to be run are identified by filename extensions,
by default \code{c(".Rmd", ".Rnw")}.
}
