% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_functions.R
\name{sample_machine_workbook_path}
\alias{sample_machine_workbook_path}
\title{Gives a file path to a sample machine excel workbook}
\usage{
sample_machine_workbook_path()
}
\value{
a path to a sample machine excel workbook bundled with this package
}
\description{
Gives a file path to a sample machine excel workbook
}
\examples{
sample_machine_workbook_path()
}
