% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{specify}
\alias{specify}
\title{Specify the IEA data}
\usage{
specify(BalancedIEAData, countries, country_colname = "Country")
}
\arguments{
\item{BalancedIEAData}{IEA data that have already been balanced}

\item{countries}{the countries for which specification should occur}

\item{country_colname}{The name of the country column in \code{BalancedIEAData}. Default is "Country".}
}
\value{
a data frame of specified IEA data
}
\description{
Specifies the IEA data in a way that is amenable to drake subtargets.
See \code{IEATools::specify_all()} for details.
}
\examples{
IEATools::sample_iea_data_path() \%>\%
  IEATools::load_tidy_iea_df() \%>\%
  make_balanced(countries = c("GHA", "ZAF")) \%>\%
  specify(countries = c("GHA", "ZAF"))
}
