% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storing.R
\name{stash_cache}
\alias{stash_cache}
\title{Save the drake cache to a zip file, then to \code{workflow_output_folder}}
\usage{
stash_cache(workflow_output_folder, dependency)
}
\arguments{
\item{workflow_output_folder}{The folder into which the drake cache will be saved.}

\item{dependency}{The target you want to be sure is executed before
calling this function.}
}
\value{
A logical saying whether the saving operation was successful.
}
\description{
Save the drake cache to a zip file, then to \code{workflow_output_folder}
}
