% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storing.R
\name{stash_psut}
\alias{stash_psut}
\title{Save the \code{PSUT} target in a pinboard.}
\usage{
stash_psut(workflow_releases_folder, psut)
}
\arguments{
\item{workflow_releases_folder}{The folder that contains the pinboard.}

\item{psut}{The PSUT object, the final target for this workflow.}
}
\value{
The fully-qualified path name of the PSUT file in the pinboard.
}
\description{
Save the \code{PSUT} target in a pinboard.
}
