% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_to_exergy.R
\name{sum_phi_vecs}
\alias{sum_phi_vecs}
\title{Sums phi_pf and phi_u vectors}
\usage{
sum_phi_vecs(
  phi_pf_vecs,
  phi_u_vecs,
  countries,
  phi_pf = IEATools::template_cols$phi_pf,
  phi_u = IEATools::template_cols$phi_u,
  phi_colname = IEATools::phi_constants_names$phi_colname
)
}
\arguments{
\item{phi_pf_vecs}{A data frame of phi_pf vectors}

\item{phi_u_vecs}{A data frame of phi_u vectors}

\item{countries}{The countries for which you want to perform this task.}
}
\value{
A data frame of summed phi_pf and phi_u vectors.
}
\description{
This function verifies that there are no rows in common between the
two input vectors.
}
