% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{target_names}
\alias{target_names}
\title{Drake cache target names}
\format{
A string list with 24 entries.
\describe{
\item{countries}{The string name of the countries target in the drake cache, giving the countries to be analyzed.}
\item{alloc_and_eff_couns}{The string name of the allocation and efficiency countries target in the drake cache, giving the countries from which final-to-useful allocations and efficiencies can be drawn for final-to-useful analyses. This is a superset of countries.}
\item{max_year}{The integer representing the maximum year to be analyzed.}
\item{iea_data_path}{A string representing the path to the file containing IEA extended energy balance data.}
\item{exemplar_table_path}{A string representing the path to the exemplar table.}
\item{fu_analysis_folder}{A string representing the folder containing final-to-useful analyses for various countries.}
\item{report_source_folders}{A string vector representing folders containing reports to be run as the last step of the workflow.}
\item{report_dest_folder}{A string representing the folder containing reports from the workflow.}
\item{AllIEAData}{The name of a data frame containing all IEA data read from \code{iea_data_path}.}
\item{IEAData}{A filtered version of \code{AllIEAData} containing information for only \code{countries}.}
\item{balanced_before}{A boolean indicating whether the \code{IEAData} are balanced before any further analysis. They usually are not, so this value is typically \code{FALSE}.}
\item{BalancedIEAData}{A balanced version of \code{IEAData}.}
\item{balanced_after}{Same as \code{balanced_before}, only for after balancing. This should be \code{TRUE}.}
\item{Specified}{A data frame containing specified IEA data.}
\item{PSUT_final}{A data frame containing \code{Specified} in a PSUT format.}
\item{IncompleteAllocationTables}{A data frame of final-to-useful allocation tables, one for each country. These allocation tables may be incomplete.}
\item{IncompleteEfficiencyTables}{A data frame of final-to-useful efficiency tables, one for each country. These efficiency tables may be incomplete.}
\item{ExemplarLists}{A data frame of lists of exemplar countries for each country in \code{countries}, and maybe more.}
\item{CompletedAllocationTables}{A data frame of completed final-to-useful allocation tables.}
\item{CompletedEfficiencyTables}{A data frame of completed final-to-useful efficiency tables.}
\item{AllocationGraphs}{A data frame containing allocation graphs.}
\item{report_source_paths}{A string vector containing paths to report files. These strings provide a detailed view of all the reports that \code{drake} will generate and should be all report files in \code{report_source_folders}.}
\item{reports_complete}{A boolean indicating success (\code{TRUE}) or failure (\code{FALSE}) of report generation.}
}
}
\usage{
target_names
}
\description{
A string list containing names of drake targets.
Items in the list provide default values for function arguments
throughout the \code{SEAPSUTWorkflow} package.
}
\examples{
target_names
}
\keyword{datasets}
