% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_to_useful.R
\name{add_C_mats}
\alias{add_C_mats}
\title{Add allocation matrices to a data frame}
\usage{
add_C_mats(
  psut_final,
  completed_allocation_tables,
  countries,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  .values = IEATools::template_cols$.values,
  c_source = IEATools::template_cols$c_source,
  C_Y = IEATools::template_cols$C_Y,
  C_EIOU = IEATools::template_cols$C_eiou
)
}
\arguments{
\item{psut_final}{a data frame containing PSUT matrices with final energy as the last stage.}

\item{completed_allocation_tables}{the completed allocation tables from which allocation (\code{C}) matrices should be created.
This data frame is most likely to be the \code{CompletedAllocationTables} target.}

\item{countries}{the countries for which C matrices should be formed}

\item{country, year}{See \code{IEATools::ieacols}.}

\item{.values, c_source, C_Y, C_EIOU}{See \code{IEATools::template_cols}.}
}
\value{
A data frame with \code{C_Y} and \code{C_EIOU} columns containing allocation matrices.
}
\description{
This function adds allocation matrices (\code{C_Y} and \code{C_EIOU}) to the previously-created
\code{CompletedAllocationTables} target.
}
