% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tes_tfc_functions.R
\name{calculate_all_ex_data}
\alias{calculate_all_ex_data}
\title{Create a data frame containing all aggregate energy/exergy data}
\usage{
calculate_all_ex_data(PSUT_DF)
}
\arguments{
\item{PSUT_DF}{}
}
\value{
A data frame containing energy/exergy values aggregated by total,
sector and product at each stage of the energy conversion chain.
}
\description{
Create a data frame containing all aggregate energy/exergy data
}
