% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tes_tfc_functions.R
\name{calculate_fu_ex_sector}
\alias{calculate_fu_ex_sector}
\title{Calculate total final consumption of final and useful energy by sector}
\usage{
calculate_fu_ex_sector(.sutdata, fd_sectors)
}
\arguments{
\item{.sutdata}{A data frame containing Physical Supply-Use Table (PSUT)
matrices with associated final demand sector names}

\item{fd_sectors}{A character vector of final demand sectors.}
}
\value{
A data frame containing total final and useful consumption by sector
}
\description{
Calculate the total final consumption (TFC) at the final and useful stages
(along with any additional stages) by sector.
This function first uses the uses \code{create_fd_sectors_list()} function,
with a user-supplied set of final demand sectors \code{fd_sectors}
to identify the final demand sectors desired for analysis.
The \code{Recca::finaldemand_aggregates()} function is then applied to \code{.sutdata}
data frame by sector, to calculate the total final consumption of all products
in each of the sectors supplied in \code{fd_sectors}.
}
\examples{
library(Recca)
tfc_sector <- calculate_fu_ex_sector(.sutdata = Recca::UKEnergy2000mats,
                                     fd_sectors = c("Residential"))

}
