% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tes_tfc_functions.R
\name{calculate_fu_ex_total}
\alias{calculate_fu_ex_total}
\title{Calculate total final demand of final and useful energy}
\usage{
calculate_fu_ex_total(PSUT_DF)
}
\arguments{
\item{PSUT_DF}{}
}
\value{
A data frame containing final and useful demand by total
}
\description{
Calculate total final demand of final and useful energy
}
