% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tes_tfc_functions.R
\name{calculate_p_ex_flow}
\alias{calculate_p_ex_flow}
\title{Calculate total primary energy by flow}
\usage{
calculate_p_ex_flow(PSUT_DF)
}
\arguments{
\item{PSUT_DF}{}
}
\value{

}
\description{
Calculate total primary energy by flow
}
