% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tes_tfc_functions.R
\name{calculate_p_ex_flow}
\alias{calculate_p_ex_flow}
\title{Calculate total primary energy by flow}
\usage{
calculate_p_ex_flow(
  .sutdata,
  p_industry_prefixes,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  year = IEATools::iea_cols$year,
  flow = IEATools::iea_cols$flow,
  e_product = SEAPSUTWorkflow::sea_cols$e_product,
  stage_col = SEAPSUTWorkflow::sea_cols$stage_col,
  gross_net = SEAPSUTWorkflow::sea_cols$gross_net,
  agg_by = SEAPSUTWorkflow::sea_cols$agg_by,
  p_ind_comp = SEAPSUTWorkflow::sea_cols$p_ind_comp,
  p_ind_prefix = SEAPSUTWorkflow::sea_cols$p_ind_prefix,
  ex = SEAPSUTWorkflow::sea_cols$ex,
  ex_p = SEAPSUTWorkflow::sea_cols$ex_p,
  primary = IEATools::all_stages$primary,
  all = "All"
)
}
\arguments{
\item{.sutdata}{A data frame containing Physical Supply-Use Table (PSUT)
matrices.}

\item{p_industry_prefixes}{A character vector of primary energy industry prefixes.
Usually "Resources", "Imports", and "Stock changes".}

\item{country, method, energy_type, year, flow}{See \code{IEATools::iea_cols}.}

\item{e_product, stage_col, gross_net, agg_by, p_ind_comp, p_ind_prefix, ex, ex_p}{See \code{SEAPSUTWorkflow::sea_cols}.}

\item{primary}{The string "Primary", representing the Primary stage of the energy conversion chain, see \code{IEATools::all_stages}.}

\item{all}{The string "All", indicating that one or both of \code{flow} and \code{product} were aggregated.}
}
\value{
A data frame containing aggregate primary energy/exergy data by flow
}
\description{
Calculate the total energy supply (TES) in primary energy terms by flow. This metric
was formerly called the total primary energy supply (TPES).
This function first uses the uses \code{Recca::find_p_industry_names()} function,
with a user-supplied set of primary industry prefixes \code{p_industry_prefixes}
to identify the primary industries desired for analysis.
The \code{Recca::primary_aggregates()} function is then applied to \code{.sutdata}
data frame by flow, to calculate the total energy supply across all products
for each flow.
}
\examples{
library(Recca)
total_energy_supply <- calculate_p_ex_flow(.sutdata = Recca::UKEnergy2000mats,
                                           p_industry_prefixes = c("Resources", "Imports"))

}
