% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{clean_targets}
\alias{clean_targets}
\title{Cleans drake targets}
\usage{
clean_targets(
  to_clean = SEAPSUTWorkflow::target_names[seq(which(SEAPSUTWorkflow::target_names ==
    first_cleaned_target), which(SEAPSUTWorkflow::target_names == last_cleaned_target))],
  first_cleaned_target = "CEDAData",
 
    last_cleaned_target = SEAPSUTWorkflow::target_names[[length(SEAPSUTWorkflow::target_names)]],
  path = NULL
)
}
\arguments{
\item{first_cleaned_target}{The string name of the first target to clean. Default is "CEDAData".}

\item{last_cleaned_target}{The string name of the last target to clean. Default is the last target in the workflow.}
}
\value{

}
\description{
It is often helpful to clean only a few targets.
This function allows specification of targets to be cleaned by names.
By default, this function cleans all targets \emph{after} \code{IEAData},
according to the list \code{SEAPSUTWorkflow::target_names}.
}
