% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{create_fd_sectors_list}
\alias{create_fd_sectors_list}
\title{Create a list containing final demand sectors}
\usage{
create_fd_sectors_list(fd_sectors, .sutdata)
}
\arguments{
\item{fd_sectors}{A character vector of final demand sectors.}

\item{.sutdata}{A data frame containing Physical Supply-Use Table (PSUT)
matrices with associated final demand sector names}
}
\value{
A list the length of a desired data frame containing final demand vectors
}
\description{
This function creates a list equal to the length of any data frame supplied.
It is typically used on a data frame containing Physical Supply-Use Tables (PSUT)
with the associated final demand sectors in the nested \code{Y} and \code{U_EIOU} matrices.
}
\examples{
library(Recca)
final_demand_sector_list <- create_fd_sectors_list(fd_sectors = c("Residential"),
.sutdata = Recca::UKEnergy2000mats)

}
