% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{extract_country_data}
\alias{extract_country_data}
\title{Extract specific country data}
\usage{
extract_country_data(
  AllIEAData,
  countries,
  max_year,
  country_colname = IEATools::iea_cols$country,
  year_colname = IEATools::iea_cols$year
)
}
\arguments{
\item{AllIEAData}{A data frame containing cleaned IEA extended energy balance data.}

\item{countries}{A list of 3-letter country codes for countries to be analyzed.}

\item{max_year}{The latest year you want to include in the extracted data.}

\item{country_colname, year_colname}{See \code{IEATools::iea_cols}.}
}
\value{
a data frame with the desired IEA data only
}
\description{
Data is extracted according to the \code{countries} object in a way that is amenable to drake subtargets.
\code{dplyr::filter()} does the subsetting.
}
\examples{
IEATools::sample_iea_data_path() \%>\%
  IEATools::load_tidy_iea_df() \%>\%
  extract_country_data(countries = c("ZAF"), max_year = 1999)
}
