% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_functions.R
\name{generate_eta_fu_template}
\alias{generate_eta_fu_template}
\title{Create a final-to-useful efficiencies template}
\usage{
generate_eta_fu_template(
  country,
  paths_target = "paths",
  fu_analysis_path_name = "fu_analysis_path",
  fu_analysis_file_name = paste0(country, " FU Analysis"),
  ext = ".xlsx",
  fu_allocation_tab_name = "FU Allocations",
  eta_fu_template_file_name = paste0(country, " FU etas Template"),
  overwrite = FALSE
)
}
\arguments{
\item{country}{a string of the 3-letter ISO country code}

\item{paths_target}{the name of the paths object in the drake cache. Default is "paths".}

\item{fu_analysis_path_name}{the name of the final-to-useful path member of the paths object. Default is "fu_analysis_path".}

\item{fu_analysis_file_name}{the name of the file containing the final-to-useful analysis for \code{country}.
Default is "\code{country} FU Analysis.\code{ext}".}

\item{ext}{the file name extension. Default is ".xlsx" for an Excel file.}

\item{fu_allocation_tab_name}{the name of the tab containing final-to-useful allocation information in \code{fu_analysis_file_name}.
Default is "FU Allocations".}

\item{eta_fu_template_file_name}{the name of the template file written by this function.
Default is "\code{country} FU etas Template.\code{ext}".}

\item{overwrite}{tells whether to overwrite an existing file named \code{eta_fu_template_file_name}.\code{ext} in location \code{fu_analysis_path_name}.
Default is \code{FALSE}.}
}
\value{
the path to the blank final-to-useful efficiency template
}
\description{
Create a final-to-useful efficiencies template
}
