% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_functions.R
\name{generate_fu_allocation_template}
\alias{generate_fu_allocation_template}
\title{Create a final-to-useful allocation template}
\usage{
generate_fu_allocation_template(
  country,
  data_target = "Specified",
  fu_analysis_path_target = "fu_analysis_folder",
  fu_allocation_template_file_name = paste0(country, " FU Allocations Template"),
  ext = ".xlsx",
  n_allocation_rows = 4,
  overwrite = FALSE,
  cache_path = ".drake/"
)
}
\arguments{
\item{country}{A string, the 3-letter ISO code of the country to be analyzed.}

\item{data_target}{The target in the drake cache from which the template is created. Default is "Specified".}

\item{fu_analysis_path_target}{The name of the target in the drake cache containing path to the final-to-useful analyses. Default is "fu_analysis_folder".}

\item{fu_allocation_template_file_name}{The file name for the template. Default is "\code{country} FU Allocations Template.\code{ext}".}

\item{ext}{the file name extension. Default is ".xlsx" for an Excel file.}

\item{n_allocation_rows}{the number of allocation rows per combination of sector and final energy carrier. Default is \code{4}.}

\item{overwrite}{tells whether to overwrite an existing file named \code{fu_allocation_file_name}.\code{ext} in location \code{fu_analysis_path_name}.
Default is \code{FALSE}.}

\item{cache_path}{The path to the drake cache. Default is ".drake/".}
}
\value{
the path to the blank final-to-useful allocation template
}
\description{
Create a final-to-useful allocation template
}
