% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{generate_reports}
\alias{generate_reports}
\title{Generate reports found in report directory}
\usage{
generate_reports(
  report_files,
 
    report_output_folder = drake::readd(SEAPSUTWorkflow::target_names[["report_output_folder"]])
)
}
\arguments{
\item{report_output_folder}{A string that identifies the directory into which the completed reports should be written.
Default is \verb{drake::readd(SEAPSUTWorkflow::target_names[["report_output_folder"]]}}

\item{report_file}{A vector of absolute paths to reports that need to be run, .
probably generated by \code{reports_to_generate()}.}
}
\value{

}
\description{
This function descends the directory rooted at \code{report_dir} and executes \code{knitr::knit} on all
files with \code{report_suffixes}.
}
\details{
Each report name is suffixed by date and time.
Previous reports are shuffled to a folder named \verb{<<report_name>>_prev}.
}
