% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_functions.R
\name{get_eta_filepaths}
\alias{get_eta_filepaths}
\title{Get all filepaths to machine excel files which contain a FIN_ETA sheet}
\usage{
get_eta_filepaths()
}
\value{
A list of the file paths to machine excel files containing
FIN_ETA frontsheets, and therefore usable data.
}
\description{
Get all filepaths to machine excel files which contain a FIN_ETA sheet
}
