% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{get_plan}
\alias{get_plan}
\title{Create a drake plan for societal exergy analysis}
\usage{
get_plan(
  countries,
  additional_exemplar_countries = NULL,
  max_year,
  how_far = "all_targets",
  iea_data_path,
  ceda_data_folder,
  machine_data_path,
  exemplar_table_path,
  fu_analysis_folder,
  reports_source_folders,
  reports_dest_folder
)
}
\arguments{
\item{countries}{A vector of abbreviations for countries whose energy conversion chain is to be analyzed,
such as "c('GHA', 'ZAF')".
Countries named in \code{countries} can also serve as exemplars for
final-to-useful allocations and efficiencies.}

\item{additional_exemplar_countries}{A vector of country abbreviations for which final-to-useful allocations
and efficiencies will be read.
An energy conversion chain will \emph{not} be constructed for these countries.
However, their final-to-useful allocations and efficiencies
may be used as exemplar information for the countries in \code{countries}.
Default is \code{NULL}, indicating no additional exemplars.}

\item{max_year}{The last year to be studied, typically the last year for which data are available.}

\item{how_far}{A string indicating the last target to include in the plan that is returned.
Default is "all_targets" to indicate all targets of the plan should be returned.}

\item{iea_data_path}{The path to IEA extended energy balance data in .csv format.}

\item{ceda_data_folder}{The path to the CEDA data in text file, .per, format.}

\item{machine_data_path}{The path to the machine data in .xlsx format.}

\item{exemplar_table_path}{The path to an exemplar table.}

\item{fu_analysis_folder}{The path to a folder containing final-to-useful analyses.
Sub-folders named with 3-letter country abbreviations are assumed.}

\item{reports_source_folders}{A string vector containing paths to folders of report sources, usually
\code{.Rnw} or \code{.Rmd} files.}

\item{reports_dest_folder}{The path to a folder into which reports are written.}
}
\value{
A drake plan object.
}
\description{
Creates a drake workflow for societal exergy analysis.
The caller specifies location of IEA data,
which countries should be analyzed, and
which the maximum year to be analyzed.
}
\details{
The return value is a \code{drake} plan object with the following targets:
\itemize{
\item \code{countries}: The countries to be analyzed, supplied in the \code{countries} argument.
\item \code{alloc_and_eff_couns}: The full set of countries for which final-to-useful allocations and efficiencies will be read. This is the sum of \code{countries} and \code{additional_exemplar_countries}, with duplicates removed.
\item \code{max_year}: The maximum year to be analyzed, supplied in the \code{max_year} argument.
\item \code{iea_data_path}: The path to IEA extended energy balance data, supplied in the \code{iea_data_path} argument.
\item \code{ceda_data_folder}: The path to the CEDA data, supplied in the \code{ceda_data_folder} argument.
\item \code{machine_data_path}: The path to the machine data excel files, supplied in the \code{machine_data_path} argument.
\item \code{exemplar_table_path}: The path to an exemplar table, supplied in the \code{exemplar_table_path} argument.
\item \code{fu_analysis_folder}: The path to the final-to-useful analysis folder, supplied in the \code{fu_analysis_folder} argument.
\item \code{report_output_folder}: The path to a report output folder, supplied in the \code{report_output_folder} argument.
\item \code{AllIEAData}: A data frame with all IEA extended energy balance data read from \code{iea_data_path}.
\item \code{IEAData}: A version of the \code{AllIEAData} data frame containing data for only those countries specified in \code{countries}.
\item \code{CEDAData}: A data frame containing temperature data supplied through \code{CEDATools::read_cru_cy_files}.
\item \code{AllMachineData}: A data frame containing Eta.fu and Phi.u values read through functions in \code{machine_functions.R}.
\item \code{balanced_before}: A boolean that tells where the data were balanced as received, usually a vector of \code{FALSE}, one for each country.
\item \code{BalancedIEAData}: A data frame containing balanced IEA extended energy balance data.
\item \code{balanced_after}: A boolean telling whether IEA extended energy balance data is balanced after balancing, usually a vector of \code{TRUE}, one for each country.
\item \code{OKToProceed}: \code{NULL} means everything is balanced and proceeding is OK.
\item \code{Specified}: A data frame with specified industries. See \code{IEATools::specify_all()}.
\item \code{PSUT_final}: A data frame containing PSUT matrices up to the final stage.
\item \code{ExemplarLists}: A data frame containing lists of exemplar countries on a per-country, per-year basis.
\item \code{IncompleteAllocationTables}: A data frame containing final-to-useful allocation tables.
\item \code{CompletedAllocationTables} : A data frame containing completed final-to-useful allocation tables.
\item \code{IncompleteEfficiencyTables}: A data frame containing final-to-useful efficiency tables.
\item \code{CompletedEfficiencyTables}: A data frame containing completed final-to-useful efficiency tables.
\item \code{Cmats} : A data frame containing \code{CompletedAllocationTables} in matrix form.
\item \code{EtaPhivecs} : A data frame containing final-to-useful effciency and exergy-to-energy ratio vectors.
\item \code{PSUT_useful} : A data frame containing PSUT matrices up to the useful stage.
\item \code{AggregateExergyEnergyData} : A data frame containing aggregate energy and exergy values by total, product, flow, and sector.
\item \code{AllocationGraphs} : A data frame containing allocation plots.
\item \code{NonStationaryAllocationGraphs} : A data frame containing allocation plots for non-stationary data only.
\item \code{EfficiencyGraphs} : A data frame containing final-to-useful efficiency plots.
\item \code{ExergyEnergyGraphs} : A data frame containing exergy-to-energy ratio plots.
\item \code{report_source_paths}: A string vector of paths to sources for reports.
\item \code{report_dest_paths}: A string for the path to a folder into which reports will written.
\item \code{reports_complete}: A boolean that tells whether reports were written successfully.
}

Callers can execute the plan by calling \code{drake::make(plan)}.
Results can be recovered with
\code{drake::readd(target = iea_data_path)} or similar.

Note that some targets can be read using \code{readd_by_country()}, including:
\itemize{
\item \code{AllIEAData},
\item \code{IEAData},
\item \code{BalancedIEAData},
\item \code{Specified},
\item \code{PSUT_final},
\item \code{ExemplarLists},
\item \code{IncompleteAllocationTables},
\item \code{CompletedAllocationTables},
\item \code{IncompleteEfficiencyTables}, and
\item \code{CompletedEfficiencyTables},
}
}
\examples{
get_plan(countries = c("GHA", "ZAF"),
         max_year = 1999,
         iea_data_path = "iea_path",
         ceda_data_folder = "ceda_path",
         machine_data_path = "machine_path",
         exemplar_table_path = "exemplar_path",
         fu_analysis_folder = "fu_folder",
         reports_source_folders = "reports_source_folders",
         reports_dest_folder = "reports_dest_folder")
}
\seealso{
\itemize{
\item \href{https://stackoverflow.com/questions/62140991/how-to-create-a-plan-in-a-function}{How to create a plan in a function}
\item \href{https://stackoverflow.com/questions/61220159/best-practices-for-unit-tests-on-custom-functions-for-a-drake-workflow}{Best practices for unit tests on custom functions for a drake workflow}
\item \href{https://github.com/tiernanmartin/drakepkg}{drakepkg}
\item \href{https://books.ropensci.org/drake/projects.html#workflows-as-r-packages}{Workflows as R packages}
}
}
