% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{get_plan}
\alias{get_plan}
\title{Create a drake plan for societal exergy analysis}
\usage{
get_plan(
  iea_data_path,
  fu_analysis_folder,
  countries,
  max_year,
  how_far = "all_targets"
)
}
\arguments{
\item{iea_data_path}{The path to IEA extended energy balance data in .csv format.}

\item{fu_analysis_folder}{The path to a folder containing final-to-useful analyses.}

\item{countries}{A vector of country abbreviations to be analyzed, such as "c('GHA', 'ZAF')".}

\item{max_year}{The last year to be studied, typically the last year for which data are available.}

\item{how_far}{Tells the last target to include in the plan that is returned.
Default is "all_targets" to indicate all targets of the plan should be returned.}
}
\value{
a drake plan object
}
\description{
Creates a drake workflow for societal exergy analysis.
The caller specifies location of IEA data,
which countries should be analyzed, and
which the maximum year to be analyzed.
}
\details{
The return value is a \code{drake} plan object with the following targets:
\itemize{
\item \code{iea_data_path}: The path to IEA extended energy balance data, supplied in the \code{iea_data_path} argument.
\item \code{countries}: The countries to be analyzed, supplied in the \code{countries} argument.
\item \code{max_year}: The maximum year to be analyzed, supplied in the \code{max_year} argument.
\item \code{AllIEAData}: A data frame with all IEA extended energy balance data read from \code{iea_data_path}.
\item \code{IEAData}: A version of the \code{AllIEAData} data frame containing data for only those countries specified in \code{countries}.
\item \code{balanced_before}: A boolean that tells where the data were balanced as received, usually a vector of \code{FALSE}, one for each country.
\item \code{BalancedIEAData}: A data frame containing balanced IEA extended energy balance data.
\item \code{balanced_after}: A boolean telling whether IEA extended energy balance data is balanced after balancing, usually a vector of \code{TRUE}, one for each country.
\item \code{OKToProceed}: \code{NULL} means everything is balanced and proceeding is OK.
\item \code{Specified}: A data frame with specified industries. See \code{IEATools::specify_all()}.
\item \code{PSUT_final}: A data frame containing PSUT matrices up to the final stage.
}

Callers can execute the plan by calling \code{drake::make(plan)}.
Results can be recovered with
\code{drake::readd(target = iea_data_path)} or similar.

Note that some targets can be read using \code{readd_by_country()}, including:
\itemize{
\item \code{AllIEAData},
\item \code{IEAData},
\item \code{BalancedIEAData},
\item \code{Specified}, and
\item \code{PSUT_final}.
}
}
\examples{
\dontrun{
  td <- tempdir()
  get_plan(iea_data_path = "mypath",
           fu_analysis_folder = td,
           countries = c("GHA", "ZAF"),
           max_year = 1999)
  unlink(td)
}
}
