% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exemplar_table_functions.R
\name{load_exemplar_table}
\alias{load_exemplar_table}
\title{Read an exemplar table from an Excel file}
\usage{
load_exemplar_table(
  exemplar_table_path = sample_exemplar_table_path(),
  exemplar_table_tab_name = "exemplar_table",
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  prev_names = "Prev.names"
)
}
\arguments{
\item{exemplar_table_path}{The path to the Excel file containing an exemplar table.
Default is the value of \code{sample_exemplar_table_path()}.}

\item{exemplar_table_tab_name}{The string name of the tab in the Excel file containing the exemplar table.
Default is "exemplar_table".}

\item{country}{The name of the country column to be created.
This column is filled with the current 3-letter ISO abbreviations for each country.
Default is "Country".}

\item{year}{The name of the year column to be created. Default is "Year".}

\item{prev_names}{The name of a column of previous names used for the country.
Default is "Prev.names".}
}
\value{
an exemplar table
}
\description{
An exemplar table has a column of country names
and additional metadata,
including the exemplar country and any name changes through time.
}
\details{
This function gathers (actually, \code{tidyr::pivot_longer()}) years into a Year column.
It also
}
\examples{
load_exemplar_table()
}
