% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{make_psut}
\alias{make_psut}
\title{Convert to PSUT matrices}
\usage{
make_psut(SpecifiedIEAData, countries, country_colname = "Country")
}
\arguments{
\item{SpecifiedIEAData}{A data frame that has already been specified via \code{specify()}.}

\item{countries}{The countries you want to convert to PSUT matrices.}

\item{country_colname}{The name of the country column in \code{SpecifiedIEAData}. Default is "Country".}
}
\value{
a \code{matsindf}-style data frame
}
\description{
Converts tidy IEA data to PSUT matrices in a way that is amenable to drake subtargets.
Internally, \code{IEATools::prep_psut()} does the conversion to matrices.
}
\examples{
IEATools::sample_iea_data_path() \%>\%
  IEATools::load_tidy_iea_df() \%>\%
  make_balanced(countries = c("GHA", "ZAF")) \%>\%
  specify(countries = c("GHA", "ZAF")) \%>\%
  make_psut(countries = c("GHA", "ZAF"))
}
