% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{readd_by_country}
\alias{readd_by_country}
\title{Read a subtarget based on country}
\usage{
readd_by_country(
  target,
  country,
  countries_target = SEAPSUTWorkflow::target_names$countries,
  cache_path = SEAPSUTWorkflow::cache_info$cache_path
)
}
\arguments{
\item{target}{The name of the drake target as a string. See \code{get_plan()} for names of targets.}

\item{country}{The 3-letter ISO abbreviation (a string) of the country
(or vector of 3-letter countries)
for whom \code{target} is to be readd from the drake cache.}

\item{countries_target}{See \code{SEAPSUTWorkflow::target_names}.}

\item{cache_path}{See \code{SEAPSUTWorkflow::cache_info}.}
}
\value{
the country-specific version of \code{target}
}
\description{
It is sometimes convenient to read a target out of the drake cache by country (or countries).
Technically, this process involves reading a sub-target,
because the plans created by \code{get_plan()} are all mapped by country.
But subtargets are extracted based on country index in the cache rather than country name, and
it is cumbersome to keep track of cache indices for countries.
This function de-references the indices based on the value of \code{country} and extracts
the correct subtarget.
}
