% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{specify}
\alias{specify}
\title{Specify the IEA data}
\usage{
specify(BalancedIEAData, countries)
}
\arguments{
\item{BalancedIEAData}{IEA data that have already been balanced}

\item{countries}{the countries for which specificaion should occur}
}
\value{
a data frame of specified IEA data
}
\description{
Specifies the IEA data in a way that is amenable to drake subtargets.
See \code{IEATools::specify_all()} for details.
}
